🐛 fix(changelog): fix subCommitScope default value to null
The commit includes changes in the packages/changelog/.changelogrc.js, packages/changelog/src/finalizeContext/index.ts, and packages/changelog/src/templates files. It fixes the default value of subCommitScope from an empty string to null when it is not defined in customConfig.scopeDisplayName. The change ensures that commits with no scope are appropriately sorted in commitGroups.