✨ feat(changelog): Add an option to show authors' avatars in summary template
This commit adds an option to show authors' avatars in the summary template of the changelog package. The new functionality is achieved by introducing a new handlebar partial called "summary-avatar," which contains the code that generates avatars for each author. The mainTemplate variable is updated to include this new partial and replace the {{gitUserInfo}} placeholder with the appropriate code based on the showAuthor and showAuthorAvatar options from the customConfig object.